refactor: extract logFields util and harden aperture client
